try sigan_initializer
